Renamed SpanDataImpl to TestSpanData and moved to subdir test (#1291)
